Finished radius, added 4 tests

+++ parkingspots/parking/models.py
@@ -5,11 +5,16 @@ from django.contrib.auth.models import User
 
 class ParkingSpot(models.Model):
     name = models.IntegerField("ID", unique=True)
-    lat = models.IntegerField("Latitude")
-    lon = models.IntegerField("Longitude")
+    lat = models.FloatField("Latitude")
+    lon = models.FloatField("Longitude")
     reserved = models.IntegerField("Reserved", default = 0)
 
     def __int__(self):
     	return self.name
 
-#TODO: reserved = models.BooleanField(False)
+
+# {	
+# 	"lat" : 37.7811,
+# 	"lon" : -122.3965,
+# 	"radius" : 1000
+# } Testing with this .json file structure in postman
